fix: extract config as secondary entry point to fix ng-packagr DTS resolution
Root cause: adding "frakton-ng": ["libs/frakton-ng"] to tsconfig paths
caused TypeScript to resolve the root package through the source directory
during ng-packagr builds. This broke DTS generation for unrelated entry
points (e.g. dialog) because sibling-entry path remapping no longer worked
correctly through the root package resolution.

Changes:
- Create libs/frakton-ng/config as a proper secondary entry point
  (ng-package.json with index.ts as entryFile)
- Move FktFieldErrorHandler type into config/models/ so it stays within
  the config rootDir
- Update config/index.ts to export FktFieldErrorHandler
- Update config/injection-tokens.ts to use local relative import
- Change fkt-field.component.ts import from 'frakton-ng' to 'frakton-ng/config'
- Remove "frakton-ng" root path from tsconfig.json (nothing in lib source
  needs to import from the root package anymore)
- Add "frakton-ng/config" path to tsconfig.json
